Teacher functions refactor

+++ routes/teacher_routes.js
@@ -86,24 +86,22 @@ router.get('/fetch_files/:id',async(req,res) => {
     }
     
     try{
+        const request = await pool.request().input('stringfiles',sql.NVarChar,'')
         let queryStr = '';
-        const request = await pool.request()
-            .input('stringfiles',sql.NVarChar,'')
 
-        //Post o comment
-        if(fetchMode === "0")
-            queryStr = 'SELECT IDPOST, TITULO, STRINGFILES, REMITENTE, FECHAHORA' + 
-                    ' FROM POST WHERE STRINGFILES IS NOT NULL AND TRIM(STRINGFILES) <> @stringfiles';
-        else if(fetchMode === "1")
-            queryStr = 'SELECT IDPOST, STRINGFILES, REMITENTE, FECHAHORA FROM COMENTARIO WHERE STRINGFILES != @stringfiles';
-        
-        
+        //SOLAMENTE POSTS   
+        queryStr = `SELECT p.IDPOST, p.TITULO, p.STRINGFILES, p.FECHAHORA,
+                    (a.MATRICULA + '-' + a.NOMBRE) AS REMITENTE
+                    FROM POST p 
+                    INNER JOIN ALUMNO a ON p.IDUSUARIO = a.IDUSUARIO
+                    WHERE p.STRINGFILES IS NOT NULL AND TRIM(STRINGFILES) <> @stringfiles`;
+    
         const results = await request.query(queryStr);
         return res.status(200).json(results.recordset);
     }
     catch(err){
         console.error('Ocurrio un error de consulta (Admin):', err);
-        res.status(500).json({message: 'Error interno del servidor'});
+        return res.status(500).json({message: 'Error interno del servidor'});
     }
 });
 
@@ -122,35 +120,31 @@ router.delete('/erase_files', async(req,res) => {
     try{
         //SE ASUME QUE TODOS LOS REQUISITOS YA ESTAN NO SE IGNORA NINGUNO
         let table = mode === 'fromPost' ? 'POST' : 'COMENTARIO';
+        let idColumn = mode === 'fromPost' ? 'IDPOST' : 'IDCOMENTARIO';
         let filesTarget = stringTarget.split('-');
         
+        //BORRAR ARCHIVO
         for (const file of filesTarget) {
             const filePath = path.resolve('./public/appUploads', file);
-            
-            //Verificar y borrar
-            await fs.accessSync(filePath);
-            await fs.unlinkSync(filePath)
+            if(fs.existsSync(filePath)){
+                fs.unlinkSync(filePath)
+            }
         }
 
         //YA SE BORRO ACTUALIZAR DATOS POST
         await pool.request()
             .input('idPost',sql.Int,idPost)
             .input('stringfiles',sql.NVarChar,stringTarget)
-            .query(`UPDATE ${table} SET STRINGFILES = '' WHERE IDPOST = @idPost AND STRINGFILES = @stringfiles`);
+            .query(`UPDATE ${table} 
+                    SET STRINGFILES = '' 
+                    WHERE ${idColumn} = @idPost AND STRINGFILES = @stringfiles`);
 
         return res.status(200).json({message: 'Archivo borrado'});
     }
     catch(err){
-        if(err.code === 'ENOENT'){
-            console.warn(`(ADMIN) Archivo no encontrado`);
-        } 
-        else {
-            console.error(`(ADMIN) Error al borrar ${err.message}`);
-        }
+        console.error(`(ADMIN) Error al borrar ${err.message}`);
+        return res.status(500).json({message: 'Error al borrar el archivo'});
     }
 });
-
-
-
 
 export default router;
